# docs/api.yaml
basePath: /
host: localhost:3344
info:
  description: Backend of repo https://github.com/vidorg/vid_vue
  license:
    name: MIT
    url: https://github.com/vidorg/vid_backend/blob/master/LICENSE
  termsOfService: https://github.com/vidorg
  title: vid backend
  version: '1.1'
paths:
...
  /v1/user/{uid}/video?page:
    get:
      consumes:
      - multipart/form-data
      description: 查询作者为用户的所有视频，返回分页数据
      operationId: -v1-user-{uid}-video?page-get
      parameters:
      - description: 用户id
        in: path
        name: uid
        required: true
        type: integer
      - description: 分页
        in: query
        name: page
        required: false
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"count\": 1,\n        \"page\": 1,\n    \
            \    \"data\": [\n            {\n                \"vid\": 10,\n      \
            \          \"title\": \"Video Title\",\n                \"description\"\
            : \"Video Demo Description\",\n                \"video_url\": \"\",\n\
            \                \"cover_url\": \"http://localhost:3344/raw/image/default/cover.jpg\"\
            ,\n                \"upload_time\": \"2019-12-26 14:14:04\",\n       \
            \         \"update_time\": \"2019-12-30 21:04:51\",\n                \"\
            author\": {\n                    \"uid\": 10,\n                    \"\
            username\": \"aoihosizora\",\n                    \"sex\": \"male\",\n\
            \                    \"profile\": \"Demo profile\",\n                \
            \    \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n                    \"birth_time\": \"2001-02-03\",\n             \
            \       \"authority\": \"normal\"\n                }\n            }\n\
            \        ]\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '404':
          description: '"user not found"'
      summary: 查询用户视频
      tags:
      - Video
...
  /v1/video/:
    post:
      consumes:
      - multipart/form-data
      description: 新建用户视频
      operationId: -v1-video-post
      parameters:
      - *id001
      - description: '视频标题

          *Minimum length* : 5

          *Maximum length* : 100'
        in: formData
        maxLength: 100
        minLength: 5
        name: title
        required: true
        type: string
      - description: '视频简介

          *Minimum length* : 0

          *Maximum length* : 255'
        in: formData
        maxLength: 255
        minLength: 0
        name: description
        required: true
        type: string
      - description: 视频资源链接
        in: formData
        name: video_url
        required: true
        type: string
      - description: 视频封面
        in: formData
        name: cover
        required: false
        type: file
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n       \"code\": 200,\n       \"message\": \"\
            success\",\n       \"data\": {\"vid\": 10, \"title\": \"Video Title\"\
            , \"description\": \"Video Demo Description\", \"video_url\": \"\", \"\
            cover_url\": \"http://localhost:3344/raw/image/default/cover.jpg\", \"\
            upload_time\": \"2019-12-26 14:14:04\", \"update_time\": \"2019-12-30\
            \ 21:04:51\", \"author\": {\"uid\": 10, \"username\": \"aoihosizora\"\
            , \"sex\": \"male\", \"profile\": \"Demo profile\", \"avatar_url\": \"\
            http://localhost:3344/raw/image/default/avatar.jpg\", \"birth_time\":\
            \ \"2001-02-03\", \"authority\": \"normal\"}}\n       }\n       }\n```"
        '400':
          description: '"request param error" / "request format error" / "request
            body too large" / "image type not supported" / "video resource has been
            used"'
        '401': *id002
        '500':
          description: '"image save failed" / "video insert failed"'
      security:
      - basicAuth: ''
      summary: 新建视频
      tags:
      - Video
  /v1/video/{vid}:
    delete:
      consumes:
      - multipart/form-data
      description: 删除用户视频
      operationId: -v1-video-{vid}-delete
      parameters:
      - *id001
      - description: 删除视频id
        in: path
        name: vid
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            \n}\n```"
        '400':
          description: '"request param error"'
        '401': *id002
        '404':
          description: '"video not found"'
        '500':
          description: '"video delete failed"'
      security:
      - basicAuth: ''
      summary: 删除视频
      tags:
      - Video
    get:
      consumes:
      - multipart/form-data
      description: 查询视频信息
      operationId: -v1-video-{vid}-get
      parameters:
      - description: 视频id
        in: path
        name: vid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n       \"code\": 200,\n       \"message\": \"\
            success\",\n       \"data\": {\"vid\": 10, \"title\": \"Video Title\"\
            , \"description\": \"Video Demo Description\", \"video_url\": \"\", \"\
            cover_url\": \"http://localhost:3344/raw/image/default/cover.jpg\", \"\
            upload_time\": \"2019-12-26 14:14:04\", \"update_time\": \"2019-12-30\
            \ 21:04:51\", \"author\": {\"uid\": 10, \"username\": \"aoihosizora\"\
            , \"sex\": \"male\", \"profile\": \"Demo profile\", \"avatar_url\": \"\
            http://localhost:3344/raw/image/default/avatar.jpg\", \"birth_time\":\
            \ \"2001-02-03\", \"authority\": \"normal\"}}\n       }\n       }\n```"
        '400':
          description: '"request param error"'
        '404':
          description: '"video not found"'
      summary: 查询视频
      tags:
      - Video
    post:
      consumes:
      - multipart/form-data
      description: 更新用户视频信息
      operationId: -v1-video-{vid}-post
      parameters:
      - *id001
      - description: 更新视频id
        in: path
        name: vid
        required: true
        type: string
      - description: '视频标题

          *Minimum length* : 5

          *Maximum length* : 100'
        in: formData
        maxLength: 100
        minLength: 5
        name: title
        required: false
        type: string
      - description: '视频简介

          *Minimum length* : 0

          *Maximum length* : 255'
        in: formData
        maxLength: 255
        minLength: 0
        name: description
        required: false
        type: string
      - description: 视频封面
        in: formData
        name: cover
        required: false
        type: file
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"vid\": 10,\n        \"title\": \"Video Title\"\
            ,\n        \"description\": \"Video Demo Description\",\n        \"video_url\"\
            : \"\",\n        \"cover_url\": \"http://localhost:3344/raw/image/default/cover.jpg\"\
            ,\n        \"upload_time\": \"2019-12-26 14:14:04\",\n        \"update_time\"\
            : \"2019-12-30 21:04:51\",\n        \"author\": {\n            \"uid\"\
            : 10,\n            \"username\": \"aoihosizora\",\n            \"sex\"\
            : \"male\",\n            \"profile\": \"Demo profile\",\n            \"\
            avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\",\n\
            \            \"birth_time\": \"2001-02-03\",\n            \"authority\"\
            : \"normal\"\n        }\n    }\n}\n```"
        '400':
          description: '"request param error" / "request format error" / "request
            body too large" / "image type not supported"'
        '401': *id002
        '404':
          description: '"video not found"'
        '500':
          description: '"image save failed" / "video update failed"'
      security:
      - basicAuth: ''
      summary: 更新视频
      tags:
      - Video
  /v1/video?page:
    get:
      consumes:
      - multipart/form-data
      description: 管理员查询所有视频，返回分页数据，Admin
      operationId: -v1-video?page-get
      parameters:
      - *id001
      - description: 分页
        in: query
        name: page
        required: false
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"count\": 1,\n        \"page\": 1,\n    \
            \    \"data\": [\n            {\n                \"vid\": 10,\n      \
            \          \"title\": \"Video Title\",\n                \"description\"\
            : \"Video Demo Description\",\n                \"video_url\": \"\",\n\
            \                \"cover_url\": \"http://localhost:3344/raw/image/default/cover.jpg\"\
            ,\n                \"upload_time\": \"2019-12-26 14:14:04\",\n       \
            \         \"update_time\": \"2019-12-30 21:04:51\",\n                \"\
            author\": {\n                    \"uid\": 10,\n                    \"\
            username\": \"aoihosizora\",\n                    \"sex\": \"male\",\n\
            \                    \"profile\": \"Demo profile\",\n                \
            \    \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n                    \"birth_time\": \"2001-02-03\",\n             \
            \       \"authority\": \"normal\"\n                }\n            }\n\
            \        ]\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '401': *id002
      security:
      - basicAuth: ''
      summary: 查询所有视频
      tags:
      - Video
      - Administration
swagger: '2.0'